[feat] HomeController 생성 #29
- HomeController.java : 초기 삭제되지 않은 데이터 만 보여주는 홈 컨트롤러 생성했습니다.
- CategoryRepository.java : 삭제되지 않은 카테고리의 전체 리스트를 가져오는 메서드 구현했습니다.
